+++ cvsnt/cvsnt-2.5.05.3744/keyValueServer/include/blob_client_lib.h
@@ -15,17 +15,17 @@ int64_t blob_pull_from_server(intptr_t &client, const char *hash_type, const cha
 
 KVRet blob_start_pull_from_server(intptr_t &sockfd, const char *hash_type, const char *hash_hex_str,
   uint64_t &from, uint64_t &sz);
-int64_t blob_pull_some_from_server(intptr_t &sockfd, char *data, size_t data_capacity, int64_t &szLeft);//if <= 0, socket closed. Should not be called when szLeft = 0
+int64_t blob_pull_some_from_server(intptr_t &sockfd, char *data, uint64_t data_capacity, int64_t &szLeft);//if <= 0, socket closed. Should not be called when szLeft = 0
 
-int64_t blob_push_to_server(intptr_t &client, size_t blob_sz,
+int64_t blob_push_to_server(intptr_t &client, uint64_t blob_sz,
   const char *hash_type, const char *hash_hex_str,
-  std::function<const char*(uint64_t at, size_t &data_pulled)> pull_data);//<0 if error. pull_data returns nullptr on error
+  std::function<const char*(uint64_t at, uint64_t &data_pulled)> pull_data);//<0 if error. pull_data returns nullptr on error
 
 KVRet blob_stream_to_server(intptr_t &client,
   const char *hash_type, const char *hash_hex_str,
-  std::function<const char*(size_t &data_pulled)> stream_data);//stream_data returns nullptr on error, data_pulled = 0 when finished
+  std::function<const char*(uint64_t &data_pulled)> stream_data);//stream_data returns nullptr on error, data_pulled = 0 when finished
 
 class StreamToServerData;
 StreamToServerData *start_blob_stream_to_server(intptr_t &sockfd, const char *hash_type, const char *hash_hex_str);
-KVRet blob_stream_to_server(StreamToServerData &strm, const void *data, size_t data_size);
+KVRet blob_stream_to_server(StreamToServerData &strm, const void *data, uint64_t data_size);
 KVRet finish_blob_stream_to_server(intptr_t &sockfd, StreamToServerData *s, bool ok);//will delete the pointer
